+++ src/classes/NativeLoader.ts
@@ -5,14 +5,16 @@ import { NativeModules } from 'react-native';
 export interface NativeLdkNodeRn {
   createConfig(
     storageDirPath: string,
+    logDirPath: string,
     network: string,
-    listeningAddress: string | null,
+    listeningAddress: Array<string> | null,
     defaultCltvExpiryDelta: number,
     onchainWalletSyncIntervalSecs: number,
     walletSyncIntervalSecs: number,
     feeRateCacheUpdateIntervalSecs: number,
-    logLevel: string,
-    trustedPeers0conf: Array<string>
+    trustedPeers0conf: Array<string>,
+    probingLiquidityLimitMultiplier: number,
+    logLevel: string
   ): string;
 
   fromConfig(configId: string): string;
@@ -24,14 +26,14 @@ export interface NativeLdkNodeRn {
   setGossipSourceRgs(buildId: string, rgsServerUrl: string): boolean;
   setStorageDirPath(buildId: string, storageDirPath: string): boolean;
   setNetwork(buildId: string, network: string): boolean;
-  setListeningAddress(buildId: string, listeningAddress: string): boolean;
+  setListeningAddresses(buildId: string, listeningAddresses: Array<string>): boolean;
   build(buildId: string): string;
 
   start(nodeId: string): boolean;
   stop(nodeId: string): boolean;
   syncWallets(nodeId: string): boolean;
   nodeId(nodeId: string): string;
-  listeningAddress(nodeId: string): string;
+  listeningAddresses(nodeId: string): Array<string> | null;
   newOnchainAddress(nodeId: string): string;
   sendToOnchainAddress(nodeId: string, address: string, amountMsat: number): string;
   sendAllToOnchainAddress(nodeId: string, address: string): string;